requirements.txt
tweepy
feedparser
requests
schedule
google-generativeai
beautifulsoup4
Pillow